refactor:전시관 관리  작품 type 정리
